moved the add friends code to dashboard.js
