+++ client/src/pages/PublicProfile.js
@@ -1,10 +1,10 @@
-import React, {useState} from 'react';
+import React from 'react';
 import { useParams } from 'react-router-dom';
 import { Link } from 'react-router-dom';
-import fuzzySearch from 'fz-search';
+
 
 import { useQuery, useMutation } from '@apollo/client';
-import { QUERY_USER, QUERY_USERS} from '../utils/queries';
+import { QUERY_USER } from '../utils/queries';
 import { ADD_FRIEND } from '../utils/mutations';
 import Auth from '../utils/auth';
 
@@ -17,12 +17,6 @@ const PublicProfile = props => {
         variables: { username: userParam }
     });
 
-    const [value, setValue] = useState("");
-    const [fuzzyValue, setFuzzyValue] = useState("");
-
-    const { loading: queryUsersLoading, data: usersData} = useQuery(QUERY_USERS);
-    console.log(usersData);
-
     if (loading) {
         return <div>Loading...</div>;
     }
@@ -34,19 +28,11 @@ const PublicProfile = props => {
     // ) {
     //     return "Welcome to your public profile page!"
     // }
-    const handleInputOnChange = e => {
-        setValue(e.target.value) 
-        //console.log(usersData)
-        const usernames = usersData.users.map(user => user.username)
 
-        const searcher = new fuzzySearch({source: usersData.users, keys:["username"]});
-        setFuzzyValue(searcher.search(e.target.value))
-    }
-
-    const handleClick = async (username) => {
+    const handleClick = async () => {
         try {
             await addFriend({
-                variables: {username}
+                variables: { id: user._id }
             });
         } catch (e) {
             console.error(e);
@@ -73,21 +59,9 @@ const PublicProfile = props => {
                         })}
                 </ul>
                 {userParam && (
-                    // <button className="btn ml-auto" onClick={handleClick}>
-                    //     Add Friend
-                    // </button>
-
-                    <div> 
-                        <p>You currently have x friends</p> 
-                        <input value={value} onChange={handleInputOnChange} type="text" />
-                         
-                        {(fuzzyValue.length > 0 && <div>
-                        {fuzzyValue.map(user => <button className="btn ml-auto" onClick={() => handleClick(user.username)}>
-                         Add {user.username}
-                     </button>)}
-                        </div>)} 
-                        
-                    </div>
+                    <button className="btn ml-auto" onClick={handleClick}>
+                        Add Friend
+                    </button>
                 )}
             </div>
